se dividen los inputs de usuario y pet en 2 componentes distintos
